feat: implementa tratamento de excecoes

+++ src/main/java/com/leandro/dscommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.leandro.dscommerce.services;
 
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.leandro.dscommerce.dto.ProductDTO;
 import com.leandro.dscommerce.entities.Product;
 import com.leandro.dscommerce.repositories.ProductRepository;
+import com.leandro.dscommerce.services.exceptions.DatabaseException;
+import com.leandro.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -22,8 +26,8 @@ public class ProductService {
 	// readOnly = true sinaliza que esse endpoint so faz consulta no banco para economizar memoria
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> result = productRepository.findById(id);
-		Product product = result.get();
+		Product product =  productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado."));
+		//Product product = result.get();
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
@@ -49,19 +53,33 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		// pega a referencia do Product pelo id passado
-		Product product = productRepository.getReferenceById(id);
-		// atualiza o product com os dados do dto
-		copyDtoToEntity(dto, product);
-		// salva no banco o novo product
-		product = productRepository.save(product);
-		// retorna o dto
-		return new ProductDTO(product);
+		try {
+			// pega a referencia do Product pelo id passado
+			Product product = productRepository.getReferenceById(id);
+			// atualiza o product com os dados do dto
+			copyDtoToEntity(dto, product);
+			// salva no banco o novo product
+			product = productRepository.save(product);
+			// retorna o dto
+			return new ProductDTO(product);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			productRepository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial!");
+		}
+		
 	}
 	
 	private void copyDtoToEntity(ProductDTO dto, Product product) {
